# routes/api.yaml
swagger: "2.0"
info:
  description: "This will serve as the official API documentation for Cockpit."
  version: "2.0.0"
  title: "COCKPIT"
  termsOfService: ""
  contact:
    email: "juan.navarrete@meggitt.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost5000"
basePath: "/v1"
...
schemes:
  - "https"
  - "http"
paths:
...
  /api/trainings:
    post:
      tags:
        - "trainings/"
      summary: "To post a new Training object"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "create a new training object"
      operationId: "post"
      parameters:
        - name: "course_title"
          in: "body"
          description: "name of the title"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/training"
      responses:
        201:
          description: "object created"
        400:
          description: "error"
    get:
      tags:
        - "trainings/"
      summary: "To post a new Training object"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "get training objects"
      operationId: "get training"
      parameters:
        - name: "course_title"
          in: "body"
          description: "name of the title"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/training"
      responses:
        200:
          description: "ok"
        500:
          description: "Internal server error"
...
  /api/visitors:
    post:
      tags:
        - "visitors"
      summary: "Create visitor"
      description: "Visitor object."
      operationId: "create Visitor"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created visitorr object"
          required: true
          schema:
            $ref: "#/definitions/visitor"
      responses:
        201:
          description: "New user created"
        400:
          description: "Error"
    get:
      tags:
        - "visitors"
      summary: "get visitors"
      description: "Get Visitors"
      operationId: "get visitors"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created visitorr object"
          required: false
          schema:
            $ref: "#/definitions/visitor"
      responses:
        200:
          description: "ok"
        500:
          description: "Error"
...
  /api/sd029:
    post:
      tags:
        - "api/sd029"
      summary: "Create a new SD029"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created Post object"
          required: true
          schema:
            $ref: "#/definitions/SD029"
      responses:
        200:
          description: "ok"
        404:
          description: "not found"
        500:
          description: "Error"
    get:
      tags:
        - "api/sd029"
      summary: "get SD029"
      description: "Get SD029"
      operationId: "get SD029"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Get posts"
          required: false
          schema:
            $ref: "#/definitions/SD029"
      responses:
        200:
          description: "ok"
        500:
          description: "Error"
  /api/sd029/id:
    get:
      tags:
        - "api/sd029/:id"
      summary: "To get SD029 object by id"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "get SD029 object by id"
      operationId: "get SD029 by _id"
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: string
      responses:
        200:
          description: "ok"
        404:
          description: "not found"
        500:
          description: "internal server error"
    patch:
      tags:
        - "api/sd029/:id"
      summary: "To update SD029 object by id"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "update sd029 object by id"
      operationId: "updateSD029"
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: string
      responses:
        200:
          description: "ok"
        404:
          description: "not found"
        500:
          description: "internal server error"
    delete:
      tags:
        - "api/sd029/:id"
      summary: "To delete SD029 object by id"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "to delete object by id"
      operationId: "delete SD029"
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: string
      responses:
        200:
          description: "ok"
        404:
          description: "not found"
        500:
          description: "internal server error"
...
definitions:
...
  visitor:
    type: "object"
    properties:
      id:
        type: "string"
        format: "moongose.object.Id()"
      name:
        type: "string"
      type:
        type: "string"
        enum:
          - "Customer Visit"
          - "Divisonal Visit"
          - "Government Visit"
          - "Offisite-Visiting Customer"
          - "Offisite-Visiting Supplier"
          - "PLC Visit"
          - "Supplier Visit"
      customer_program:
        type: "string"
      attendees:
        type: string
      meeting_location:
        type: string
      start_date:
        type: string
        format: date-time
      end_date:
        type: string
        format: date-time
      purpose:
        type: string
  training:
    type: "object"
    properties:
      id:
        type: "string"
        format: "moongose.object.Id()"
      course_title:
        type: "string"
      training_type:
        type: "string"
      date:
        type: "string"
        format: date-time
      instructor:
        type: "string"
      number_of_session:
        type: "string"
      hours_per_class:
        type: "number"
      employee_population:
        type: "number"
...
  SD029:
    type: "object"
    properties:
      id:
        type: string
        format: "moongose.object.Id()"
      part_number:
        type: string
      invc_dt:
        type: string
        format: date-time
      so_num:
        type: number
      so_ln:
        type: number
      so_typ:
        type: string
      qty:
        type: number
      lc_tp:
        type: number
      matlGrp:
        type: string
      invc_num:
        type: number
      part_description:
        type: string
      lc_up:
        type: number
      dest_regn:
        type: string
      ship_to_name:
        type: string
      ship_to_num:
        type: number
      dest_ctry:
        type: string
      dc:
        type: number
      sdv:
        type: number
      cg:
        type: string
      mdv:
        type: number
      mg1:
        type: string
      mg5:
        type: string
      ph1:
        type: string
      ph2:
        type: string
      ph3:
        type: string
      ph4:
        type: string
      ph5:
        type: string
      ph4_desc:
        type: string
      sold_to_name:
        type: string
      customer_number:
        type: string
      invdt_fiscyp:
        type: number
      suom:
        type: string
      lc_tc:
        type: number
      lc_uc:
        type: number
      lc_gm_amt:
        type: number
      gm_percent:
        type: number
      lcurr:
        type: string
      exch_rt:
        type: number
      dcurr:
        type: string
      dc_tp:
        type: number
      dc_up:
        type: number
      dc_tc:
        type: number
      dc_uc:
        type: number
      dc_gm_amt:
        type: number
      sofc:
        type: string
      cust_matl_num:
        type: string
      profit_ctr:
        type: number
      mrp_contr:
        type: string
      mrp_contr_nm:
        type: string
      proctyp:
        type: string
      mrp_grp:
        type: string
      cust_po:
        type: string
      cust_po_date:
        type: string
        format: date-time
      uommul:
        type: number
      uomdiv:
        type: number
      buom:
        type: string
      sls_mgr_num:
        type: number
      sls_mgr_nm:
        type: string
      emp_resp_number:
        type: number
      emp_resp_nm:
        type: string
      cust_grp1:
        type: string
      cust_grp_1_desc:
        type: string
      sold_ctry:
        type: string
      dlvprio:
        type: number
      ext_agt_number:
        type: string
      ext_agt_nm:
        type: string
      end_usr_number:
        type: number
      end_usr_nm:
        type: string
      end_usr_ctry:
        type: string
      sldv:
        type: string
      invc_yr:
        type: number
      wbs:
        type: string
      mlstn:
        type: number
      ph6:
        type: number
      ha00:
        type: number
      yk08:
        type: number
      yfrt:
        type: number
      slcrtdt:
        type: string
        format: date-time
      cust_grp_desc:
        type: string
      prodschd:
        type: string
      ship_to_zc:
        type: string
      so_incot_1:
        type: string
      so_inco_desc:
        type: string
      so_incot_2:
        type: string
      so_payt:
        type: string
      so_payt_desc:
        type: string
      del_num:
        type: number
      invc_typ:
        type: string
      ord_resn:
        type: string
      ord_resn_desc:
        type: string
      invoice_text:
        type: string
      ship_to_ctry:
        type: string
      bill_to_num:
        type: string
      bill_to_nam:
        type: string
      primcontrno:
        type: string
      cascvd:
        type: string
      tina:
        type: string
      gup:
        type: string
      gup_name:
        type: string
      sales_org:
        type: number
      plant:
        type: number
...